web/pyproject.toml
[tool.black]
line-length = 110
exclude = "migrations"
skip-string-normalization = true

[tool.pytest.ini_options]
norecursedirs = [
  ".git ",
  "node_modules",
]
python_files = "*test*.py"
python_classes = "*Test*"
DJANGO_SETTINGS_MODULE = "src.settings_prod"
log_cli = "true"
log_cli_level = "info"


[tool.coverage.run]
omit = [
  "*/migrations/*",
  "src/settings*",
  "manage.py",
  "*/apps.py",
]

[tool.coverage.report]
exclude_lines = [
  "^\\s*@(abc.)?abstractmethod",
  "^\\s*@(typing.)?overload",
  "^\\s*if (typing.)?TYPE_CHECKING:",
  "^\\s*if (settings.)?DEBUG:",
  "pragma: no ?cover",
  "def __repr__",
  "def __str__",
  "if self.debug:",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
]

[tool.coverage.xml]
output = "coverage.xml"

[tool.mypy]
python_version = "3.10"
cache_dir = ".cache/mypy"
exclude = [
  "tests"
]
disallow_untyped_defs = true
plugins = [
  "mypy_django_plugin.main",
  "mypy_drf_plugin.main"
]

[[tool.mypy.overrides]]
module = ["*.migrations.*", "manage"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ['celery.*', 'django_filters', 'kombu.*', 'drf_yasg.*', 'factory.*', 'src.additional_settings.*']
ignore_missing_imports = true


[tool.django-stubs]
django_settings_module = "src.settings_prod"
ignore_missing_model_attributes = true

[tool.doc8]
max_line_length = 110

[tool.isort]
multi_line_output = 3
skip = "migrations"
line_length = 110
include_trailing_comma = true

[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Горелов Михаил <gorelovmihail0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.13"
psycopg2-binary = "2.9.3"
redis = "3.5.3"
djangorestframework = "3.13.1"
drf-yasg = "1.20.0"
celery = "5.2.6"
PyAMQP = "0.1.0.7"
django-filter = "21.1"
django-cors-headers = "3.11.0"
django-defender = "0.9.3"
django-redis = "5.2.0"
django-rosetta = "0.9.8"
sentry-sdk = "1.5.10"
django-microservice-request = "0.5.1"
Pillow = "8.1.1"
django-countries = "7.3.2"
django-silk = "4.3.0"
coverage = {version = "6.3.2", extras = ["toml"]}
freezegun = "1.2.1"
pipdeptree = "2.2.1"
django-debug-toolbar = "3.2.4"
django-debug-toolbar-request-history = "0.1.4"
black = "22.3.0"
flake8 = "4.0.1"
isort = "5.10.1"
mypy = "0.942"
django-stubs = "1.10.1"
djangorestframework-stubs = "1.5.0"
pytest = "7.1.2"
pytest-mock = "3.7.0"
pytest-cov = "3.0.0"
pytest-django = "4.5.2"
requests-mock = "1.9.3"
gunicorn = "20.1.0"
uvicorn = {version = "0.17.6", extras = ["standard"]}
django-allauth = "0.50.0"
dj-rest-auth = "2.2.4"
djangorestframework-simplejwt = "5.1.0"
django-cacheops = "6.0"
jsonschema = "4.4.0"
django-treebeard = "4.5.1"
Unidecode = "1.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
